### 2018-04-25 #705 @whedon
[PRE REVIEW]: egtplot: A Python Package for Three-Strategy Evolutionary Games
**Submitting author:** @drew-williamson (<a href="http://orcid.org/0000-0003-1745-8846">Drew Williamson</a>)
**Repository:** <a href="https://github.com/mirzaevinom/egtplot" target ="_blank">https://github.com/mirzaevinom/egtplot</a>
**Version:** v0.3
