### 2018-04-25 0654d1e
figure edits
