### 2018-04-25 #705 @drew-williamson
I've updated `paper.md` in the repo so that the figures should now (fingers crossed) display correctly.
