### 2018-04-25 #705 @drew-williamson
Forgot to add figure 2 to the repo. Once more?
