### 2018-04-26 #705 @arfon
> Forgot to add figure 2 to the repo. Once more? Thanks!

You can ask Whedon to do this for you :-)
