### 2018-04-26 b3f0940
figure sizing
