+++ paper.md
@@ -23,6 +23,12 @@ affiliations:
   index: 3
 date: 25 April 2018
 bibliography: joss_bib.bib
+
+output:
+  rmarkdown::pdf_document:
+    fig_caption: yes        
+    includes:  
+      in_header: figure_placement.tex
 ---
 
 # Summary
@@ -59,11 +65,12 @@ Using our package, we can quickly and easily analyze this game numerically and v
 
 By altering the default values for the plotting function, a variety of different plotting styles can be achieved as demonstrated in __Figure 2__.
 
-![(_Left_) Background of the simplex is colored by the speed at which the points would travel along their trajectories. (_Right_) Displaying the paths taken by each initial condition.](images/two_parts.png){ width=50% }
+![(_Left_) Background of the simplex is colored by the speed at which the points would travel along their trajectories. (_Right_) Displaying the paths taken by each initial condition.](images/two_parts.png){ width=70% }
 
 Additionally, multiple parameter values can be easily combined into subplots of a larger image. In __Figure 3__, we vary $\alpha$ and $\beta$ values to see how they would independently affect the dynamics of the game.
 
 ![Parameter sweep for different $\alpha$ and $\beta$ values.](images/output_20_1.png){ width=50% }
+
 
 Finally, `egtplot` also has functionality to display animated versions of these plots. For further information on the use of the software and how these plots were created, please see the [documentation](https://github.com/mirzaevinom/egtplot) and [example notebook](https://github.com/mirzaevinom/egtplot/blob/master/egtplot_demonstration.ipynb).
 
