+++ float_adjustment.tex
@@ -1,0 +1,2 @@
+\usepackage{float}
+\floatplacement{figure}{H}
