### 2018-04-27 dbf7671
fixed equilibria
