### 2018-05-06 252d4d7
major equilibrium update
