+++ egtplot/__init__.py
@@ -3,6 +3,7 @@ from __future__ import division
 import inspect
 import sys
 from itertools import product
+import copy
 
 import matplotlib.gridspec as gridspec
 import matplotlib.patches as patches
@@ -31,7 +32,7 @@ def random_ics(ic_num):
     # draw points from the unit cube uniformly at random
     points = np.random.random((ic_num, 3))
 
-    # ensure the sum of each point's coordinates sum to 1
+    # ensure the sum of each point's coordinates is 1
     # (i.e, the point lies in the simplex)
     total = np.sum(points, axis=1).reshape((-1, 1))
     ics = np.divide(points, total)
@@ -179,150 +180,293 @@ def grid_ics():
     return ics
 
 
-# plot equilibria
-def equilibria(payoffs, ax):
-    """ Equilibria that are a single point will be plotted as such, with an
-    open circle indicating instability, filled circle indicating stability,
-    and a grey circle indicating sem-stability (i.e. that the second derivative
-    is exactly zero. When every initial condition is at an equilibrium in a
-    subgame, that is plotted as a cross-hatched rectangle over the entire edge.
+def proj_to_from(x, y):
+    """
+    Project points to 3D and back to 2D inside the simplex
+    """
+    return tuple(np.dot(proj, np.array([1 / (1 + x + y), x / (1 + x + y), y / (1 + x + y)])))
 
-    It's important to note that these equilibria are for points that lie
-    EXACTLY on the edge only."""
+
+def get_extraploted_line(p1,p2, EXTRAPOL_RATIO = 10):
+    """
+    Creates a line extrapoled in both directions p1->p2 and p2->p1
+    """
+
+    a = (p2[0]+EXTRAPOL_RATIO*(p1[0]-p2[0]), p2[1]+EXTRAPOL_RATIO*(p1[1]-p2[1]) )
+    b = (p1[0]+EXTRAPOL_RATIO*(p2[0]-p1[0]), p1[1]+EXTRAPOL_RATIO*(p2[1]-p1[1]) )
+
+    return LineString([a, b])
+
+
+def load_bomze_payoffs():
+    """
+    Loads Bomze payoffs as a list
+    """
+    import pickle
+    with open('./bomze_payoffs.pkl', 'rb') as input_file:
+        bomze_payoffs = pickle.load(input_file)
+
+    return bomze_payoffs
+
+
+def equilibria(payoffs, ax):
+
+    # copy the payoffs so we can alter one version and still have the original
+    payoff_mat = copy.deepcopy(payoffs)
+
+    # prepare for Bomze's testing (Appendix of Biol Cyb, 1983)
+    # transform payoffs to [0 0 0], [a b c], [d e f]
+    payoffs[:, 0] -= payoffs[0, 0]
+    payoffs[:, 1] -= payoffs[0, 1]
+    payoffs[:, 2] -= payoffs[0, 2]
+
+    # define variables for testing following Bomze's conventions
+    a = payoffs[1, 0]
+    b = payoffs[1, 1]
+    c = payoffs[1, 2]
+    d = payoffs[2, 0]
+    e = payoffs[2, 1]
+    f = payoffs[2, 2]
+
+    i = b * f - c * e
+    j = a * e - b * d
+    k = c * d - a * f
+
+    ####################
+    # domain equilibria
+    ####################
+    line_indicator = 0
+
+    if np.allclose(payoffs, np.zeros((3, 3))):
+        pass
+
+    elif i == 0 and j == 0 and k == 0:
+        line_indicator = 1
+
+        if c != 0:
[…]
