### 2018-05-06 a49634b
updated readme with equilibria
