### 2018-05-07 d287e94
fixed load_bomze_payoffs function
