### 2018-05-14 #705 @arfon
@nagoue @jmbr - would you be willing to review this submission for JOSS?
