### 2018-05-15 #705 @arfon
@fil - would you be willing to review this submission for JOSS? Ideally I'd like to have two reviewers for this submission (@jmbr has already kindly volunteered)
