### 2018-05-15 #705 @Fil
yes — will remind me of my research 25 yrs ago that was on this topic :)
