### 2018-05-16 mirzaevinom/egtplot#1 @jmbr
Minor revisions for JOSS
Here are my suggestions for egtplot:

1. Split `__init__.py` into two files: one for the proper use of `__init.py__` (see https://docs.python.org/3/tutorial/modules.html#packages for details) and another one containing the actual code. It is uncommon and a bit perplexing to place the whole code for a package in the init file (see the Python documentation linked above).
2. The version declared in setup.py (0.4.4) does not match the version reported by JOSS (0.3).
3. In the notebook file, there are a number of typos:
- Institue -> Institute
- artibitrary -> arbitrary
- display_paramters -> display_parameters
- equilirbia -> equilibria
- equilirbium -> equilibrium
- micronenvironmentally -> microenvironmentally
- mulitple -> multiple
- protate -> prostate
- provideo -> provide
- qualititative -> qualitative
4. There is an indirect reliance on the Python requests module through the MoviePy package. Before installing requests I got the following traceback:
```
<ipython-input-15-00e5eefebdf8> in <module>()
      1 parameter_values = [[1], [1], [1], [1]]
----> 2 animation, fps = plot_animated(parameter_values, custom_func=get_payoff)
      3 animation.write_gif('animation_1.gif', program='imageio', loop=1, fps=fps)

~/tmp/egtplot/egtplot/__init__.py in plot_animated(payoff_entries, num_fps, generations, steps, ic_type, ic_num, ic_dist, dot_color, dot_size, eq, display_parameters, custom_func, vert_labels)
    735     import imageio
    736     imageio.plugins.ffmpeg.download()
--> 737     from moviepy.editor import VideoClip
    738     from moviepy.video.io.bindings import mplfig_to_npimage
    739

/anaconda3/envs/egtplot/lib/python3.6/site-packages/moviepy/editor.py in <module>()
     29 from .video.io.VideoFileClip import VideoFileClip
     30 from .video.io.ImageSequenceClip import ImageSequenceClip
---> 31 from .video.io.downloader import download_webfile
     32 from .video.VideoClip import VideoClip, ImageClip, ColorClip, TextClip
     33 from .video.compositing.CompositeVideoClip import CompositeVideoClip, clips_array

/anaconda3/envs/egtplot/lib/python3.6/site-packages/moviepy/video/io/downloader.py in <module>()
      5 import os
      6
----> 7 import requests
      8
      9 from moviepy.tools import subprocess_call

ModuleNotFoundError: No module named 'requests'

```
5. It would be desirable to have automated unit tests.
6. Although the API is documented in the Jupyter notebook file, it would be better if the module also had comprehensive docstrings for `plot_static` and `plot_animated`.
7. Despite the fact that the code is available on Github, the JOSS review guidelines demand that the package provides instructions on how to contribute to the software, report issues, or seek support.
