### 2018-05-16 #735 @jmbr
Accept with minor revisions: https://github.com/mirzaevinom/egtplot/issues/1
