### 2018-05-16 #735 @drew-williamson
Thanks so much for taking the time to review egtplot! This is our first Python package so it's a bit rough around the edges. We'll make the revisions right away.
