### 2018-05-18 #735 @Fil
So far I haven't been able to install and run the demo notebook. I have struggled with incompatibilities between versions of macos, conda, geopandas, tornado, jupyter etc. (I wanted to try it in an virtual environment.)

I keep hitting those errors:
https://github.com/jupyter/notebook/issues/3407
https://github.com/geopandas/geopandas/issues/556

I might try later but not sure I'll find time.
