### 2018-05-21 mirzaevinom/egtplot#2 @Fil
Issues for JOSS
Review for JOSS (https://github.com/openjournals/joss-reviews/issues/735)

- **Installation**

I struggled to install the package on anaconda/macos, but finally managed to install everything on a docker container and the notebook is now running smoothly in jupyter lab.

I used:
```
pip install matplotlib scipy shapely imageio moviepy tqdm jupyter
```

I then added `pip install requests` to make the movie export work. But for some reason it still gets stuck at 99%
```
[MoviePy] Building file animation_2.gif with imageio
 99%|█████████▉| 120/121 [00:05<00:00, 21.12it/s]
```

other than that, it seems to work fine, creating the expected diagrams.

It would be good to make sure installation is easier — maybe a complete Dockerfile could help some users.

Regarding:
- ** A statement of need **: it is clear in the paper, but not in the README.md
- ** Automated tests** . and ** Community guidelines** would be welcome
