### 2018-05-21 #735 @Fil
I finally managed to install almost everything with docker.

Here are my comments so far: https://github.com/mirzaevinom/egtplot/issues/2
