### 2018-05-28 f5b2031
JOSS review #1 edits
