+++ egtplot/__init__.py
@@ -1,914 +1,1 @@
-from __future__ import division
-
-import inspect
-import sys
-from itertools import product
-import copy, os
-
-import matplotlib.gridspec as gridspec
-import matplotlib.patches as patches
-import matplotlib.pyplot as plt
-import numpy as np
-from scipy.integrate import odeint
-from scipy.spatial.distance import cdist
-from shapely.geometry import Polygon, LineString, Point
-from tqdm import tqdm
-
-# define the projection to triangular coordinates
-proj = np.array([[-1 * np.cos(30. / 360. * 2. * np.pi),
-                  np.cos(30. / 360. * 2. * np.pi), 0.],
-                 [-1 * np.sin(30. / 360. * 2. * np.pi),
-                  -1 * np.sin(30. / 360. * 2. * np.pi),
-                  1.]])
-
-# define the vertices and edges of the simplex
-trianglepoints = np.hstack([np.identity(3), np.array([[1.], [0.], [0.]])])
-triangleline = np.dot(proj, trianglepoints)
-
-
-def random_ics(ic_num):
-    # generate initial conditions randomly distributed within the simplex
-
-    # draw points from the unit cube uniformly at random
-    points = np.random.random((ic_num, 3))
-
-    # ensure the sum of each point's coordinates is 1
-    # (i.e, the point lies in the simplex)
-    total = np.sum(points, axis=1).reshape((-1, 1))
-    ics = np.divide(points, total)
-
-    return ics
-
-
-def random_uniform_points(ic_num, ic_dist):
-    # generate randomly distributed points inside the simplex
-    # separated from each other by a distance.
-
-    # draw points from the unit cube uniformly at random
-    ics = np.random.uniform(low=0.0, high=1.0, size=[ic_num, 3])
-
-    # ensure the sum of each point's coordinates sum to 1
-    # (i.e, the point lies in the simplex)
-    ics = (ics.T / np.sum(ics, axis=1)).T
-
-    # project the points
-    tri_cond = np.dot(proj, np.array(ics).T).T
-
-    # compute the distances between all points
-    distances = cdist(tri_cond, tri_cond)
-    size = len(tri_cond)
-    distances = np.tril(distances) + np.triu(np.ones((size, size)))
-
-    # get rid of all points that are within a distance from each other
-    indices = np.unique(np.nonzero(np.min(distances, axis=1) > ic_dist)[0])
-    ics = ics[indices]
-
-    # generate points along each edge
-    n_edge = 10
-    xy_edge = 0.0 * np.ones([n_edge, 3])
-    xy_edge[:, 0] = np.linspace(0, 1, n_edge, endpoint=False)
-    xy_edge[:, 1] = 1 - xy_edge[:, 0]
-
-    yz_edge = 0.0 * np.ones([n_edge, 3])
-    yz_edge[:, 0] = np.linspace(0, 1, n_edge, endpoint=False)
-    yz_edge[:, 2] = 1 - yz_edge[:, 0]
-
-    xz_edge = 0.0 * np.ones([n_edge, 3])
-    xz_edge[:, 1] = np.linspace(0, 1, n_edge, endpoint=False)
-    xz_edge[:, 2] = 1 - xz_edge[:, 1]
-
-    # shove all the points together then get rid of duplicates
-    ics = np.concatenate((ics, xy_edge, yz_edge, xz_edge), axis=0)
-    ics = np.unique(ics, axis=1)
-
-    return ics
-
-
-def max_num_ics(min_dist, num_per_side=10):
-    """This is a helper function to determine approximately how many initial
-    conditions will be returned by a call to random_initial conditions without
-    actually calling the function itself. For large values of ic_num, the total
-    assymtotically approaches 1.3 divided by the area of the disk centered at
[…]
