### 2018-05-28 20abf07
JOSS review #2 edits
