### 2018-05-28 dd74c4d
JOSS review edits
