+++ readme.md
@@ -9,7 +9,7 @@ This is a software package for plotting and animating three-strategy evolutionar
 
 <img src="images/animation_1.gif" width="400" height="400" />
 
-## Use Cases
+### Use Cases
 
 While there are many use cases for this software, our group's particular focus is on the mathematical modeling of cancer and cancer therapies through evolutionary game theory [@kaznatcheev2017cancer, @kaznatcheev2015edge, @Kaznatcheev179259]. For details on analytical treatments of evolutionary games, please see Artem Kaznatcheev's blog at [Theory, Evolution, and Games Group](https://egtheory.wordpress.com/).
 
@@ -50,7 +50,7 @@ cd egtplot-master
 python setup.py install
 ```
 
-### Usage
+### Specific Usage
 * This software has two main functions: `plot_static` for plotting static simplex figures and `plot_animated` for generating simplex animations.
 * For detailed usage of these functions we refer to our interactive jupyter notebook: [`egtplot_demonstration.ipynb`](egtplot_demonstration.ipynb)
 * We also welcome comments and questions regarding our whitepaper on [bioRxiv](https://www.biorxiv.org/content/early/2018/04/12/300004) which describes the package and its usage. The content is nearly identical to the jupyter notebook linked above.
