### 2018-05-28 mirzaevinom/egtplot#1 @drew-williamson
Thanks for these, @jmbr! @mirzaevinom and I have (we think) addressed all of your points, with the exception of (2). We haven't been able to figure out how to change the version in JOSS. Do you know or should we ask one of the editors?
