### 2018-05-28 mirzaevinom/egtplot#1 @jmbr
Excellent. I recommend you ask the editors how to change the declared version number.
