### 2018-05-28 #735 @drew-williamson
@mirzaevinom and I have made the edits requested by @jmbr and are now working on those requested by @Fil. A question for @arfon, though: how do we change the version of the package on JOSS?
