### 2018-05-29 #735 @arfon
> @mirzaevinom and I have made the edits requested by @jmbr and are now working on those requested by @Fil. A question for @arfon, though: how do we change the version of the package on JOSS?

@drew-williamson - I can change the version. What will the new version number be?
