### 2018-05-29 f712e0b
"bug fixes and JOSS review #2"
