+++ egtplot/egtplot_code.py
@@ -229,7 +229,7 @@ def load_bomze_payoffs():
         import urllib
 
         url = (
-            "https://raw.github.com/mirzaevinom/egtplot/master/egtplot/bomze_payoffs.pkl"
+          "https://raw.github.com/mirzaevinom/egtplot/master/egtplot/bomze_payoffs.pkl"
         )
         urllib.request.urlretrieve(url, fname)
 
@@ -452,7 +452,7 @@ def equilibria(payoffs, ax):
             eq_subgame_full[0][subgame[1]] += 1 - eq_subgame
             eqs.append(eq_subgame_full)
 
-            # make a point to test numerically that's close to the equilibrium but is in the domain
+            # make a point to test that's close to the equilibrium but is in the domain
             moved = eq_subgame_full.copy()
             ind = np.argmin(moved)
             moved[0, ind] += 0.01
@@ -637,28 +637,56 @@ def plot_static(
 ):
     """Function to plot static evolutionary game solutions.
 
-    Arguments:
-        payoff_entries (list): list of nine lists containing entries in the payoff matrix
-        generations (int): the number of epochs to simulate forward the ODEs; default 6.
-        steps (int): the number of steps simulated per generation; default 200
-        background (bool): controls whether the background of the simplex is colored by speed; default False
-        ic_type (str): the distribution of initial conditions ('random', 'grid', or 'edge'); default 'grid'
-        ic_num (int): the number of initial conditions; default 100
-        ic_dist (float): the distance between initial conditions when ic_type = 'random'; default 0.05
-        ic_color (string): the color of each initial condition; default 'black'
-        paths (bool): whether the paths taken by each initial condition should be displayed; default False
-        path_colors (string): the colormap used for the paths for each initial condition; default 'inferno'
-        eq (bool): whether or not equilibria should be plotted on the simplex; default True
-        display_parameters (bool): whether or not to print the payoff matrix next to the simplex; default True
-        custom_func (function): makes the payoffs a function of other parameters; default None
-        vert_labels (list): list of strings that should label the vertices; default ['X', 'Y', 'Z']
+    Arguments
+    ---------
+        payoff_entries (list): list of nine lists containing entries in the payoff
+            matrices
 
-    Returns:
+        generations (int): the number of epochs to simulate forward the ODEs; default 6
+
+        steps (int): the number of steps simulated per generation; default 200
+
+        background (bool): controls whether the background of the simplex is colored by
+            speed; default False
+
+        ic_type (str): the distribution of initial conditions ('random', 'grid', or
+            'edge'); default 'grid'
+
+        ic_num (int): the number of initial conditions; default 100
+
+        ic_dist (float): the distance between initial conditions when ic_type =
+            'random'; default 0.05
+
+        ic_color (string): the color of each initial condition; default 'black'
+
+        paths (bool): whether the paths taken by each initial condition should be
+            displayed; default False
+
+        path_colors (string): the colormap used for the paths for each initial
+            condition; default 'inferno'
+
+        eq (bool): whether or not equilibria should be plotted on the simplex;
+            default True
+
+        display_parameters (bool): whether or not to print the payoff matrix next to
+            the simplex; default True
+
+        custom_func (function): makes the payoffs a function of other parameters;
+            default None
+
+        vert_labels (list): list of strings that should label the vertices; default
+            ['X', 'Y', 'Z']
+
+    Returns
+    -------
         A matplotlib figure object containing the designated simplex.
 
-    Please see https://github.com/mirzaevinom/egtplot/blob/master/egtplot_demonstration.ipynb for more.
+    More
+    ----
+    See https://github.com/mirzaevinom/egtplot/blob/master/egtplot_demonstration.ipynb
+    for greater detail an examples.
[…]
