### 2018-05-29 mirzaevinom/egtplot#2 @drew-williamson
Thank you @Fil for the comments. We have created a docker container located at https://hub.docker.com/r/drewfkw/egtplot/ to ease installation. Sorry yours was so troublesome!

A statement of need has been added to `README.md`, automated tests have been added in the file `test_egtplot.py`, and community guidelines have been added as a section to `README.md` entitled "Contributing and Issues."
