### 2018-05-29 mirzaevinom/egtplot#2 @Fil
The docker image installs nicely, and the demo notebook is easy to find - thanks a lot for this.

The demo notebook runs well, but it stops at the exact same spot as it did on my <del>local</del> <ins>own docker</ins> installation.

However I now understand 💡 that the kernel has probably saved the gif somewhere on disk — it just doesn't show up where it's supposed to, in the next cell (and I was troubled by 120/121 which seems incomplete… but the cell had finished).

And, applying the `animation.ipython_display(fps=fps, loop=True, autoplay=True)` command suggested a few lines later works perfectly.

Once this is resolved (in any way you wish), I'm good.
