### 2018-05-29 mirzaevinom/egtplot#2 @drew-williamson
We have fixed the gif loading error. The issue was simply a lack of a . before /images in the html loading the gifs. Not sure why tqdm stops at 99% and 120/121 though the cell finishes running--very strange.

Thanks!
