### 2018-06-15 #735 @jmbr
@arfon I closed the issue https://github.com/mirzaevinom/egtplot/issues/1 some weeks ago. It's good to go as far as I am concerned.
